+++ AeSys/EoGsModelTransform.h
@@ -3,7 +3,7 @@
 #include "EoGeMatrix3d.h"
 
 class EoGsModelTransform {
-	size_t m_Depth;
+	unsigned m_Depth;
 
 	EoGeMatrix3d m_CurrentModelTransform;
 	EoGeMatrix3dList m_TransformMatrixList;
@@ -13,7 +13,7 @@ public:
 	EoGsModelTransform();
 	~EoGsModelTransform();
 	EoGeMatrix3d ModelMatrix() const noexcept;
-	size_t Depth() const noexcept;
+	unsigned Depth() const noexcept;
 	/// <summary> Removes the top transformation off the current transformation stack.</summary>
 	void PopModelTransform();
 	/// <summary> The specified transformation is concatenated to the current model transformation (which is initially the identity transform).</summary>
